+++ lib/pay/lemon_squeezy/webhooks/payment.rb
@@ -1,0 +1,11 @@
+module Pay
+  module LemonSqueezy
+    module Webhooks
+      class Payment
+        def call(event)
+          Pay::LemonSqueezy::Charge.sync(event.data.id)
+        end
+      end
+    end
+  end
+end